+++ src/ahkDebug.ts
@@ -178,7 +178,10 @@ export class AhkDebugSession extends LoggingDebugSession {
         runtimeArgs,
         {
           cwd: path.dirname(args.program),
-          env: args.env,
+          env: {
+            ...process.env,
+            ...args.env,
+          },
         },
       );
       ahkProcess.on('close', (exitCode) => {