+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Also want to check the development status, check the [commit history](https://gi
 ## [1.6.2] - 2020-10-xx
 ### Changed
 * [#56](https://github.com/zero-plusplus/vscode-autohotkey-debug/issues/56) Error messages are now output in UTF-8 whenever possible
+* Improve error message when an invalid value is set by launch.json
 
 ### Fixed
 * [#63](https://github.com/zero-plusplus/vscode-autohotkey-debug/issues/63) Can't get environment variables in the script